allow for 64 bit copy types to take the space of the inner pointer

+++ src/vm/core_types/mod.rs
@@ -1,6 +1,6 @@
 use crate::vm::prelude::*;
 
-impl CanObj for i128 {
+impl CanObj for i64 {
     fn get_item(obj: &Obj<PtyPtr>, key: &str) -> Obj<PtyPtr> {
         todo!()
     }
